src/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ado-copilot-agent"
version = "0.1.0"
description = "Azure DevOps work item lifecycle automation using GitHub Copilot CLI"
requires-python = ">=3.12"
dependencies = [
    "typer>=0.21.1",
    "rich>=13.0.0",
    "python-dotenv>=1.2.1",
    "requests>=2.31.0",
    "gitpython>=3.1.40",
]

[project.scripts]
ado-copilot-agent = "cli:cli"

[tool.setuptools]
packages = ["commands", "services", "models", "utilities"]
py-modules = ["cli"]

[dependency-groups]
dev = [
    "pyinstaller>=6.0.0",
    "pytest>=8.0",
    "pytest-mock>=3.0",
]
